+++ kobart/__init__.py
@@ -21,9 +21,7 @@
 # TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 # OR OTHER DEALINGS IN THE SOFTWARE.
 
-from kobart import utils
-from kobart import pytorch_kobart
-from kobart.utils import get_kobart_tokenizer, download
-from kobart.pytorch_kobart import get_pytorch_kobart_model
+from kobart.utils import download
+from kobart.pytorch_kobart import get_pytorch_kobart_model, get_kobart_tokenizer
 
 __all__ = ("download", "get_kobart_tokenizer", "get_pytorch_kobart_model")
